+++ include/stumpless/target.h
@@ -116,26 +116,6 @@ struct stumpless_target {
 int stump( const char *message, ... );
 
 /**
- * Logs a message to the default target.
- *
- * This is an alias for the stump function, which is preferred. This function
- * will be deprecated in a future release.
- *
- * @param message The message to log, optionally containing any format
- * specifiers valid in \c printf.
- *
- * @param ... Substitutions for any format specifiers provided in message. The
- * number of substitutions provided must exactly match the number of
- * specifiers given.
- *
- * @return A non-negative value if no error is encountered. If an error is
- * encountered, then a negative value is returned and an error code is set
- * appropriately.
- */
-int
-stumpless( const char *message, ... );
-
-/**
  * Logs a message to the default target with the given priority.
  *
  * This function can serve as a replacement for the traditional \c syslog
@@ -458,28 +438,6 @@ stumpless_unset_option( struct stumpless_target *target, int option );
  */
 int
 vstump( const char *message, va_list subs );
-
-/**
- * Logs a message to the default target.
- *
- * This is an alias for the vstump function, which is preferred. This function
- * will be deprecated in a future release.
- *
- * @param message The message to log, optionally containing any format
- * specifiers valid in \c printf.
- *
- * @param subs Substitutions for any format specifiers provided in message. The
- * number of substitutions provided must exactly match the number of
- * specifiers given. This list must be started via \c va_start before being
- * used, and \c va_end should be called afterwards, as this function does not
- * call it.
- *
- * @return A non-negative value if no error is encountered. If an error is
- * encountered, then a negative value is returned and an error code is set
- * appropriately.
- */
-int
-vstumpless( const char *message, va_list subs );
 
 /**
  * Logs a message to the default target with the given priority. Can serve as