+++ test/function/cpp/param.cpp
@@ -20,7 +20,7 @@
 #include <gtest/gtest.h>
 #include <stumpless.hpp>
 
-using namespace stumplesscpp;
+using namespace stumpless;
 
 namespace {